[Improv-14083][Resource Center] Handle deleteSource in OSS / S3 / GCS (#14084)

+++ dolphinscheduler-storage-plugin/dolphinscheduler-storage-gcs/src/main/java/org/apache/dolphinscheduler/plugin/storage/gcs/GcsStorageOperator.java
@@ -144,7 +144,7 @@ public class GcsStorageOperator implements Closeable, StorageOperate {
     }
 
     @Override
-    public void download(String tenantCode, String srcFilePath, String dstFilePath, boolean deleteSource,
+    public void download(String tenantCode, String srcFilePath, String dstFilePath,
                          boolean overwrite) throws IOException {
         File dstFile = new File(dstFilePath);
         if (dstFile.isDirectory()) {
@@ -201,7 +201,9 @@ public class GcsStorageOperator implements Closeable, StorageOperate {
                         .setTarget(target)
                         .build());
 
-        gcsStorage.delete(source);
+        if (deleteSource) {
+            gcsStorage.delete(source);
+        }
         return true;
     }
 
@@ -212,7 +214,12 @@ public class GcsStorageOperator implements Closeable, StorageOperate {
             BlobInfo blobInfo = BlobInfo.newBuilder(
                     BlobId.of(bucketName, dstPath)).build();
 
-            gcsStorage.create(blobInfo, Files.readAllBytes(Paths.get(srcFile)));
+            Path srcPath = Paths.get(srcFile);
+            gcsStorage.create(blobInfo, Files.readAllBytes(srcPath));
+
+            if (deleteSource) {
+                Files.delete(srcPath);
+            }
             return true;
         } catch (Exception e) {
             log.error("upload failed,the bucketName is {},the filePath is {}", bucketName, dstPath);
